% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolomicsWorkbenchParam.R, R/MsExperiment.R
\name{MwbParam}
\alias{MwbParam}
\alias{readMsObject,MsExperiment,MwbParam-method}
\title{Load Content from a Metabolomics Workbench Study}
\usage{
MwbParam(
  mwbId = character(),
  analysisId = character(),
  filePattern = "mzML$|CDF$|cdf$|mzXML$"
)

\S4method{readMsObject}{MsExperiment,MwbParam}(object, param, ...)
}
\arguments{
\item{mwbId}{\code{character(1)} The Metabolomics Workbench study ID, which should
start with "ST". This identifier uniquely specifies the study within
the Metabolomics Workbench database.}

\item{analysisId}{\code{character(1)} The Metabolomics Workbench analysis ID,
which should start with "AN". This identifier uniquely specifies the
analysis within the Metabolomics Workbench database.}

\item{filePattern}{\code{character(1)} A regular expression pattern to filter the
raw data files associated with the selected assay. The default value is
\code{"mzML$|CDF$|cdf$|mzXML$"}, corresponding to the supported raw data file
types.}

\item{object}{For \code{readMsObject()}: a \code{MsExperiment} instance.}

\item{param}{For \code{readMsObject()}: a \code{MwbParam} object.}

\item{...}{Currently ignored.}
}
\value{
\code{readMsObject()} returns an \code{MsExperiment} object with the
\code{sampleData()} populated with Metabolomics Workbench sample information
and the experiment's MS data loaded as a \link[Spectra:Spectra]{Spectra::Spectra} object.
}
\description{
The \code{MwbParam} class and the associated \code{readMsObject()} method
allow users to load an \link[MsExperiment:MsExperiment]{MsExperiment::MsExperiment} object from a study in
the Metabolomics Workbench database (https://metabolomicsworkbench.org/) by
providing its unique study identifier (parameter \code{mwbId}). This function
is particularly useful for directly importing metabolomics data (MS data and
experimental metadata) into an \code{MsExperiment} object for further analysis
in the R environment.

It is important to note that at present it is only possible to \emph{read}
(import) data from Metabolomics Workbench, but not to \emph{save} data to
Metabolomics Workbench.

If the study contains multiple analyses, the user will be prompted to select
which analysis to load. The resulting \code{MsExperiment} object will include a
\code{sampleData} slot populated with data extracted from the selected analysis.

The resulting \code{MsExperiment} object will include a \code{sampleData} slot
populated with data extracted from the Metabolomics Workbench metadata.

Further filtering can be performed using the \code{filePattern} parameter of the
\code{MwbParam} object. The default for this parameter is
\code{"mzML$|CDF$|cdf$|mzXML$"}, which corresponds to the supported raw data file
types.
}
\examples{

library(MsExperiment)
## Load a study with the mwbId "ST002115" and selecting specific file
## pattern as well as removing ontology and protocol information in the
## metadata.
param <- MwbParam(mwbId = "ST002115", analysisId = "AN003513",
                  filePattern = "01_RP.mzXML")
ms_experiment <- readMsObject(MsExperiment(), param)
ms_experiment

## The object's sampleData contains information loaded from Metabolomics
## Workbench
sampleData(ms_experiment)

## The MS data files were downloaded and cached; the data is available
## through the object's `Spectra`
spectra(ms_experiment)
}
\seealso{
\itemize{
\item \link[MsExperiment:MsExperiment]{MsExperiment::MsExperiment} object.
\item \link[MsBackendMetabolomicsWorkbench:MsBackendMetabolomicsWorkbench]{MsBackendMetabolomicsWorkbench::MsBackendMetabolomicsWorkbench} for
retrieving MS data files from Metabolomics Workbench
\item \href{https://metabolomicsworkbench.org/}{Metabolomics Workbench} for accessing
the Metabolomics Workbench database.
}
}
\author{
Gabriele Tomè
}
