% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboLightsParam.R, R/MsExperiment.R
\name{MetaboLightsParam}
\alias{MetaboLightsParam}
\alias{readMsObject,MsExperiment,MetaboLightsParam-method}
\title{Load Content from a MetaboLights Study}
\usage{
MetaboLightsParam(
  mtblsId = character(),
  assayName = character(),
  filePattern = "mzML$|CDF$|cdf$|mzXML$"
)

\S4method{readMsObject}{MsExperiment,MetaboLightsParam}(
  object,
  param,
  keepOntology = TRUE,
  keepProtocol = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{mtblsId}{\code{character(1)} The MetaboLights study ID, which should
start with "MTBL". This identifier uniquely specifies the study within
the MetaboLights database.}

\item{assayName}{\code{character(1)} The name of the assay to load. If the study
contains multiple assays and this parameter is not specified, the user
will be prompted to select which assay to load.}

\item{filePattern}{\code{character(1)} A regular expression pattern to filter the
raw data files associated with the selected assay. The default value is
\code{"mzML$|CDF$|cdf$|mzXML$"}, corresponding to the supported raw data file
types.}

\item{object}{For \code{readMsObject()}: a \code{MsExperiment} instance.}

\item{param}{For \code{readMsObject()}: a \code{MetaboLightsParam} object.}

\item{keepOntology}{\code{logical(1)} Whether to keep columns related to ontology
in the object's \code{sampleData()}. Default is \code{TRUE}.}

\item{keepProtocol}{\code{logical(1)} Whether to keep columns related to protocols
information in the object's \code{sampleData()}. Default is \code{TRUE}.}

\item{simplify}{\code{logical(1)} Whether to simplify the \code{sampleData()} table by
removing columns filled with NAs or duplicated content. Default is
\code{TRUE}.}

\item{...}{Currently ignored.}
}
\value{
\code{readMsObject()} returns an \code{MsExperiment} object with the
\code{sampleData()} populated with MetaboLights sample and assay information
and the experiment's MS data loaded as a \link[Spectra:Spectra]{Spectra::Spectra} object.
}
\description{
The \code{MetaboLightsParam} class and the associated \code{readMsObject()} method
allow users to load an \link[MsExperiment:MsExperiment]{MsExperiment::MsExperiment} object from a study in
the MetaboLights database (https://www.ebi.ac.uk/metabolights/index) by
providing its unique study identifier (parameter \code{mtblsId}). This function
is particularly useful for directly importing metabolomics data into an
\code{MsExperiment} object for further analysis in the R environment.

It is important to note that at present it is only possible to \emph{read}
(import) data from MetaboLights, but not to \emph{save} data to MetaboLights.

If the study contains multiple assays (e.g. measurements performed in
positive or negative polarity or data subsets with different liquid
chromatography setups used), the user will be prompted to select
which assay to load. The resulting \code{MsExperiment} object will include a
\code{sampleData} slot populated with data extracted from the selected assay.

Users can define how to filter this \code{sampleData} table by specifying a few
parameters. The \code{keepOntology} parameter is set to \code{TRUE} by default, meaning
that all ontology-related columns are retained. If set to \code{FALSE}, they are
removed. If ontology columns are kept, some column names may be duplicated
and therefore numbered. The order of these columns is important, as it
reflects the assay and sample information available in MetaboLights.

The \code{keepProtocol} parameter is also set to \code{TRUE} by default, meaning that
all columns related to protocols are kept. If set to \code{FALSE}, they are
removed. The \code{simplify} parameter (default \code{simplify = TRUE}) allows to
define whether duplicated columns or columns containing only missing values
should be removed. In the case of duplicated content, only the first
occurring column will be retained.

Further filtering can be performed using the \code{filePattern} parameter of the
\code{MetaboLightsParam} object. The default for this parameter is
\code{"mzML$|CDF$|cdf$|mzXML$"}, which corresponds to the supported raw data file
types.
}
\examples{

library(MsExperiment)
## Load a study with the mtblsId "MTBLS39" and selecting specific file
## pattern as well as removing ontology and protocol information in the
## metadata.
param <- MetaboLightsParam(mtblsId = "MTBLS39", filePattern = "63A.cdf")
ms_experiment <- readMsObject(MsExperiment(), param, keepOntology = FALSE,
                              keepProtocol = FALSE)
ms_experiment

## The object's sampleData contains information loaded from MetaboLights
sampleData(ms_experiment)

## The MS data files were downloaded and cached; the data is available
## through the object's `Spectra`
spectra(ms_experiment)
}
\seealso{
\itemize{
\item \link[MsExperiment:MsExperiment]{MsExperiment::MsExperiment} object.
\item \link[MsBackendMetaboLights:MsBackendMetaboLights]{MsBackendMetaboLights::MsBackendMetaboLights} for retrieving MS data
files from MetaboLights.
\item \href{https://www.ebi.ac.uk/metabolights/index}{MetaboLights} for accessing
the MetaboLights database.
}
}
\author{
Philippine Louail
}
