% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperimentFiles.R
\name{MsExperimentFilesStash}
\alias{MsExperimentFilesStash}
\alias{saveObject,MsExperimentFiles-method}
\alias{saveMsObject,MsExperimentFiles,AlabasterParam-method}
\alias{readMsObject,MsExperimentFiles,AlabasterParam-method}
\title{Stash for \code{MsExperimentFiles}}
\usage{
\S4method{saveObject}{MsExperimentFiles}(x, path, ...)

\S4method{saveMsObject}{MsExperimentFiles,AlabasterParam}(object, param, ...)

\S4method{readMsObject}{MsExperimentFiles,AlabasterParam}(object, param, ...)
}
\arguments{
\item{x}{An \code{MsExperimentFiles} object.}

\item{path}{For \code{saveObject()}: \code{character(1)} with the path where the
object should be stored into.}

\item{...}{Currently ignored.}

\item{object}{An \code{MsExperimentFiles} object.}

\item{param}{An \code{AlabasterParam}.}
}
\value{
\code{readMsObject()} returns a \link[MsExperiment:MsExperimentFiles]{MsExperiment::MsExperimentFiles} object.
}
\description{
The \link[MsExperiment:MsExperimentFiles]{MsExperiment::MsExperimentFiles} class stores files (or rather file
names) that are part of a mass spectrometry experiment.

The supported stash formats for \code{MsExperimentFiles} objects are listed in
the sections below.
}
\section{\emph{alabaster}-based format, \code{AlabasterParam}}{


The \code{MsExperimentFiles} stash folder contains the alabaster-specific
\emph{OBJECT} file and a sub-folder \emph{x} with the \code{MsExperimentFiles}
content serialized by \emph{alabaster.base}.
}

\examples{

library(MsExperiment)

fls <- MsExperimentFiles(list(input = c("file.mzML", "file2.mgf")))

## Define the path to the stash
d <- file.path(tempdir(), "ms_file_stash")

## Stash the object in alabaster format
saveMsObject(fls, AlabasterParam(d))

## The content of the stash: subfolder x contains the *character list*
## saved through the *alabaster.base* package.
library(fs)
dir_tree(d)

## Restore the object from stash
res <- readMsObject(MsExperimentFiles(), AlabasterParam(d))
res

## In addition, it is possible to read the object also with the
## *alabaster.base* functionality
library(alabaster.base)
res <- readObject(d)
}
\author{
Johannes Rainer
}
