% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperiment.R
\name{MsExperimentStash}
\alias{MsExperimentStash}
\alias{saveObject,MsExperiment-method}
\alias{saveMsObject,MsExperiment,AlabasterParam-method}
\alias{readMsObject,MsExperiment,AlabasterParam-method}
\title{\code{MsExperiment} Stash}
\usage{
\S4method{saveObject}{MsExperiment}(x, path, ...)

\S4method{saveMsObject}{MsExperiment,AlabasterParam}(object, param, ...)

\S4method{readMsObject}{MsExperiment,AlabasterParam}(object, param, ...)
}
\arguments{
\item{x}{A \code{MsExperiment} object.}

\item{path}{For \code{saveObject()}:}

\item{...}{For \code{saveMsObject()}: optional arguments passed down to the
\code{saveMsObject()} function to stash the \code{Spectra} object (if present),
such as \code{consolidate}. For \code{readMsObject()}: optional arguments for the
\code{readMsObject()} call to restore the \code{Spectra} object (such as
\code{spectraPath}). See \link[SpectraStash:SpectraStash]{SpectraStash::SpectraStash} for more information.}

\item{object}{A \code{MsExperiment} object.}

\item{param}{The parameter object to select and configure the stash format.
Either \link[MsStash:AlabasterParam]{MsStash::AlabasterParam} or \link[MsStash:PlainTextParam]{MsStash::PlainTextParam}.}
}
\description{
\code{MsExperiment} objects can be stored to (or read from) \emph{MsExperimentStash}es
using the \code{saveMsObject()} and \code{readMsObject()} functions which take a
second argument \code{parameter} to select and configure the format of the stash.

The supported stash formats are listed in the sections below.
}
\note{
Overwriting an existing \emph{MsExperimentStash} is not allowed.

Serializing \code{MsExperiment} objects containing a \code{QFeatures} object is
currently not supported.

The \emph{plain text file-based} stash is currently not supported.
}
\section{\emph{abalbaster}-based format, \code{AlabasterParam}}{


This stash format is the most complete and reliable way for long-term
(and portable) storage of an \code{MsExperiment}. Objects can be saved or read
from this stash format either using the \code{saveMsObject()} and \code{readMsObject()}
functions or also using the \code{\link[alabaster.base:saveObject]{alabaster.base::saveObject()}} and
\code{\link[alabaster.base:readObject]{alabaster.base::readObject()}} functions.
Data from the object's slots are stored to their respective folders (using
alabaster functionality). These folders are:
\itemize{
\item \emph{experiment_files}: the content of the \verb{@experimentFiles} slot, stored
as a \link{MsExperimentFilesStash}.
\item \emph{metadata}: the content of the object's \verb{@metadata} slot.
\item \emph{other_data}: the content of the object's \verb{@otherData} slot. Note that
export fails if object types are stored in this slot without available
\emph{alabaster} export functionality.
\item \emph{qdata}: the content of the object's \verb{@qdata} slot (if present). Currently
only \link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment} objects are supported.
\item \emph{sample_data}: the object's \code{sampleData} data frame.
\item \emph{sample_data_links}: the content of the object's \verb{@sampleDataLinks} slot
defining the mapping between rows in \code{sampleData} and other entities in
the object, such as e.g. spectra.
\item \emph{sample_data_links_mcols}: the metadata content of the \verb{@sampleDataLinks}.
\item \emph{spectra}: the \link[Spectra:Spectra]{Spectra::Spectra} object with the MS data (if present).
The respective \link[SpectraStash:SpectraStash]{SpectraStash::SpectraStash} functionality is used to
export this data.
Note that not all \code{MsBackend} types might be supported. In that case,
the backend should be switched to one of the \code{MsBackend}s from the
\emph{Spectra} package using the \code{\link[Spectra:setBackend]{Spectra::setBackend()}} function.
}
}

\examples{

## Example MS data files
library(Spectra)
library(MsExperiment)
library(MsDataHub)
fls <- c(X20171016_POOL_POS_1_105.134.mzML(),
    X20171016_POOL_POS_3_105.134.mzML())

## Create a MsExperiment for the two example files
mse <- readMsExperiment(fls, data.frame(name = c("A", "B"), index = 1:2))

## Define the path where to create the MsExperimentStash
d <- file.path(tempdir(), "ms_experiment_stash")

## Save the MsExperiment to a stash in alabaster format; Note: with
## `consolidate = TRUE` the MS data files are also copied into the
## stash
saveMsObject(mse, AlabasterParam(d), consolidate = TRUE)

## Show the content of the stash folder
library(fs)
dir_tree(d)

## Restore the object from the stash
res <- readMsObject(MsExperiment(), AlabasterParam(d))
res

sampleData(res)

spectra(res)
}
\author{
Philippine Louail, Johannes Rainer
}
